HDPI-1367: tidyingup code
