+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/event/ResumePossessionClaim.java
@@ -48,14 +48,11 @@ import uk.gov.hmcts.reform.pcs.ccd.service.ClaimService;
 import uk.gov.hmcts.reform.pcs.ccd.service.PartyService;
 import uk.gov.hmcts.reform.pcs.ccd.service.PcsCaseService;
 import uk.gov.hmcts.reform.pcs.ccd.service.UnsubmittedCaseDataService;
-import uk.gov.hmcts.ccd.sdk.type.ListValue;
-import uk.gov.hmcts.reform.pcs.ccd.domain.DefendantDetails;
 import uk.gov.hmcts.reform.pcs.ccd.type.DynamicStringList;
 import uk.gov.hmcts.reform.pcs.ccd.type.DynamicStringListElement;
 import uk.gov.hmcts.reform.pcs.postcodecourt.model.LegislativeCountry;
 import uk.gov.hmcts.reform.pcs.security.SecurityContextService;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
@@ -177,7 +174,7 @@ public class ResumePossessionClaim implements CCDConfig<PCSCase, State, UserRole
         String contactEmail = isNotBlank(pcsCase.getOverriddenClaimantContactEmail())
             ? pcsCase.getOverriddenClaimantContactEmail() : pcsCase.getClaimantContactEmail();
 
-        populateDefendantsList(pcsCase);
+        pcsCaseService.populateDefendantsList(pcsCase);
 
         PcsCaseEntity pcsCaseEntity = pcsCaseService.patchCase(caseReference, pcsCase);
         PartyEntity party = partyService.createAndLinkParty(
@@ -201,36 +198,5 @@ public class ResumePossessionClaim implements CCDConfig<PCSCase, State, UserRole
         unsubmittedCaseDataService.deleteUnsubmittedCaseData(caseReference);
     }
 
-    /**
-     * Populates the defendants list from individual defendant fields.
-     * Only processes non-null defendants and applies field clearing logic.
-     *
-     * @param pcsCase the case data containing individual defendant fields
-     */
-    private void populateDefendantsList(PCSCase pcsCase) {
-        List<ListValue<DefendantDetails>> defendantsList = new ArrayList<>();
-        DefendantDetails[] individualDefendants = {
-            pcsCase.getDefendant1(), pcsCase.getDefendant2(), pcsCase.getDefendant3(),
-            pcsCase.getDefendant4(), pcsCase.getDefendant5(), pcsCase.getDefendant6(),
-            pcsCase.getDefendant7(), pcsCase.getDefendant8(), pcsCase.getDefendant9(),
-            pcsCase.getDefendant10(), pcsCase.getDefendant11(), pcsCase.getDefendant12(),
-            pcsCase.getDefendant13(), pcsCase.getDefendant14(), pcsCase.getDefendant15(),
-            pcsCase.getDefendant16(), pcsCase.getDefendant17(), pcsCase.getDefendant18(),
-            pcsCase.getDefendant19(), pcsCase.getDefendant20(), pcsCase.getDefendant21(),
-            pcsCase.getDefendant22(), pcsCase.getDefendant23(), pcsCase.getDefendant24(),
-            pcsCase.getDefendant25()
-        };
-        
-        for (DefendantDetails defendant : individualDefendants) {
-            if (defendant != null) {
-                defendantsList.add(new ListValue<>(UUID.randomUUID().toString(), defendant));
-            }
-        }
-        
-        if (!defendantsList.isEmpty()) {
-            pcsCaseService.clearHiddenDefendantDetailsFields(defendantsList);
-            pcsCase.setDefendants(defendantsList);
-        }
-    }
 
 }
